+++ Assemble/Util/Constant/mockServerConstants.swift
@@ -1,0 +1,13 @@
+//
+//  mockServerConstants.swift
+//  Assemble
+//
+//  Created by 이창준 on 2022/04/15.
+//
+
+import Foundation
+
+enum MockServerConstants {
+    static let baseURL = "https://c34acec2-303f-4243-8b23-bd1fc8acd4c5.mock.pstmn.io"
+    static let upcomingFilmPath = "/films/upcoming"
+}
